[FIX] delete Items / improved useExpenses-Hook

+++ src/hooks/useExpenses.js
@@ -16,8 +16,8 @@ export default function useExpenses() {
             .catch(error => console.error('Fehler beim Abrufen der Daten:', error));
     }
 
-    const fetchItemsByDate = (date) => {
-        const parsedDate = new Date(date);
+    const fetchItemsByDate = async (date) => {
+        const parsedDate = await new Date(date);
         return fetch(`${url}${parsedDate.getFullYear()}/${parsedDate.getMonth() + 1}`)
             .then((res) => {
                 if (!res.ok) {
@@ -34,7 +34,7 @@ export default function useExpenses() {
             name: "",
             category: 0,
             price: 0,
-            date: new Date(date.$d)
+            date: new Date(date)
         };
 
         await fetch(url, {
@@ -48,7 +48,6 @@ export default function useExpenses() {
                 if (!response.ok) {
                     throw new Error('Fehler beim Hinzufügen einer Zeile');
                 }
-                fetchItemsByDate();
             })
             .catch(error => console.error('Fehler beim Hinzufügen einer Zeile:', error));
     }
@@ -67,35 +66,32 @@ export default function useExpenses() {
                 throw new Error('Fehler beim Aktualisieren einer Zeile');
             }
             const responseData = await response.json();
-            await fetchItemsByDate(responseData.date);
             return responseData;
-            
+
         } catch (error) {
             console.error('Fehler beim Aktualisieren einer Zeile:', error);
         }
     }
 
-
-    const deleteItems = async (selectedItems) => {
-        console.log(selectedItems);
-        /* const jsonBody = {
-            ids: selectedItems
-        };
-        await fetch(url, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(jsonBody)
-        }).then(response => {
-            if (!response.ok) {
-                throw new Error('Fehler beim Löschen einer Zeile');
-            }
-            // Aktualisieren Sie den Zustand nach erfolgreichem Hinzufügen
-            fetchItems();
-        })
-            .catch(error => console.error('Fehler beim Löschen einer Zeile:', error)); */
+    const deleteItems = (selectedItems) => {
+        if (selectedItems.length > 0) {
+            const jsonBody = {
+                ids: selectedItems
+            };
+            fetch(url, {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(jsonBody)
+            }).then(response => {
+                if (!response.ok) {
+                    throw new Error('Fehler beim Löschen einer Zeile');
+                }
+            })
+                .catch(error => console.error('Fehler beim Löschen einer Zeile:', error));
+        }
     }
 
-    return { items, fetchItemsByDate, addItem, updateItem, deleteItems};
+    return { items, fetchItemsByDate, addItem, updateItem, deleteItems };
 }
